refactor(styles): move the DodamLoading in styles package to components package
