--- packages/styles/src/DodamAnimation/DodamSkeleton.ts
@@ -1,1 +1,0 @@
-import styled from "styled-components";
