--- packages/styles/src/DodamAnimation/index.ts
@@ -1,2 +1,0 @@
-export * from "./DodamSkeleton";
-export * from "./DodamLoading";
