+++ packages/styles/src/index.ts
@@ -1,5 +1,5 @@
-export * from "./DodamAnimation";
 export * from "./DodamGlobalStyles";
 export * from "./DodamShape";
 export * from "./DodamTheme";
+export * from "./DodamSkeleton";
 export * from "./DodamTypography";
